Split into modules and sub-modules
Also introduced differential module with grad_factory() that dispatches correct grad function depending on dimension.
This can be cached in physics modules constructor.

Had to introduce a `dim` attribute for simplicity in Field and TrainableField.

TODO:
* Add tests for dim
* Add tests for differential

Ref #8
