+++ src/neurom/physics/__init__.py
@@ -1,0 +1,4 @@
+from neurom.physics.term import Term
+from neurom.physics.elastic_energy import ElasticEnergy
+from neurom.physics.load_potential import LoadPotential
+
